+++ R/create.project.R
@@ -3,18 +3,17 @@
 #' This function will create all of the scaffolding for a new project.
 #' It will set up all of the relevant directories and their initial
 #' contents. For those who only want the minimal functionality, the
-#' \code{minimal} argument can be set to \code{TRUE} to create a subset of
+#' \code{template} argument can be set to \code{minimal} to create a subset of
 #' ProjectTemplate's default directories. For those who want to dump
 #' all of ProjectTemplate's functionality into a directory for extensive
 #' customization, the \code{dump} argument can be set to \code{TRUE}.
 #'
 #' @param project.name A character vector containing the name for this new
 #'   project. Must be a valid directory name for your file system.
-#' @param minimal A boolean value indicating whether to create a minimal
-#'   project or a full project. A minimal project contains only the
-#'   directories strictly necessary to use ProjectTemplate and does not
-#'   provide template code for profiling, unit testing or documenting your
-#'   project.
+#' @param template A character vector containing the name of the template to
+#'   use for this project. By default a \code{full} and \code{minimal} template
+#'   are provided, but custom templates can be created using
+#'   \code{create.template}.
 #' @param dump A boolean value indicating whether the entire functionality
 #'   of ProjectTemplate should be written out to flat files in the current
 #'   project.
@@ -38,35 +37,22 @@
 #' library('ProjectTemplate')
 #'
 #' \dontrun{create.project('MyProject')}
-create.project <- function(project.name = 'new-project', minimal = FALSE,
+create.project <- function(project.name = 'new-project', template = 'full',
                            dump = FALSE, merge.strategy = c("require.empty", "allow.non.conflict"))
 {
-  temp.dir <- tempfile("ProjectTemplate")
-  parent_dir <- .project.info(dirname(getwd()))
-  this_dir <- .project.info(getwd())
-  if (parent_dir$is.ProjectTemplate | this_dir$is.ProjectTemplate) {
-          return(
-                  message(paste0(c(paste0("Current directory: ", basename(getwd()),
-                                          " is a ProjectTemplate directory"),
-                                   "Cannot create a new project inside an existing one",
-                                   "Please change to another directory and re-run create.project()"),
-                                 collapse = "\n")
-                  )
-          )
-  }
 
-  if (minimal) {
-    exclude <- c("diagnostics", "doc", "graphs", "lib", "logs", "profiling",
-                 "reports", "tests", "TODO")
-  } else {
-    exclude <- c()
-  }
+  .stopifproject(c("Cannot create a new project inside an existing one",
+                           "Please change to another directory and re-run create.project()"))
+
+  .stopifproject(c("Cannot create a new project inside an existing one",
+                   "Please change to another directory and re-run create.project()"),
+                 path = dirname(getwd()))
 
   merge.strategy <- match.arg(merge.strategy)
   if (.is.dir(project.name)) {
-    .create.project.existing(project.name, merge.strategy, exclude)
+    .create.project.existing(project.name, merge.strategy, template)
   } else
-    .create.project.new(project.name, exclude)
+    .create.project.new(project.name, template)
 
   if (dump)
   {
@@ -88,26 +74,25 @@ create.project <- function(project.name = 'new-project', minimal = FALSE,
   invisible(NULL)
 }
 
-.create.project.existing <- function(project.name, merge.strategy, exclude) {
-  template.path <- system.file('defaults/full', package = 'ProjectTemplate')
-  template.files.all <- .list.files.and.dirs(path = template.path)
-  template.files <- setdiff(template.files.all, exclude)
+.create.project.existing <- function(project.name, merge.strategy, template) {
+  template.path <- .get.template(template)
+  template.files <- .list.files.and.dirs(path = template.path)
 
   project.path <- file.path(project.name)
 
   switch(
     merge.strategy,
-    require.empty={
+    require.empty = {
       if (!.dir.empty(project.path))
         stop(paste("Directory", project.path,
                    "not empty.  Use merge.strategy = 'allow.non.conflict' to override."))
     },
-    allow.non.conflict={
+    allow.non.conflict = {
       target.file.exists <- file.exists(file.path(project.path, template.files))
       if (any(target.file.exists))
         stop(paste("Creating a project in ", project.path,
                    " would overwrite the following existing files/directories:\n",
-                   paste(template.files[target.file.exists], collapse=', ')))
+                   paste(template.files[target.file.exists], collapse = ', ')))
     },
     stop("Invalid value for merge.strategy:", merge.strategy))
 
@@ -121,7 +106,7 @@ create.project <- function(project.name = 'new-project', minimal = FALSE,
   writeLines(c(sprintf("# %s\n", basename(normalizePath(project.name))), README), README.md)
 }
 
-.create.project.new <- function(project.name, exclude) {
+.create.project.new <- function(project.name, template) {
   if (file.exists(project.name)) {
     stop(paste("Cannot run create.project() from a directory containing", project.name))
   }
@@ -130,16 +115,13 @@ create.project <- function(project.name = 'new-project', minimal = FALSE,
   tryCatch(
     .create.project.existing(project.name = project.name,
                              merge.strategy = "require.empty",
-                             exclude = exclude),
+                             template = template),
     error = function(e) {
       unlink(project.name, recursive = TRUE)
       stop(e)
     }
   )
 }
-
-.get.template.tar.path <- function(template.name)
-  system.file(file.path('defaults', paste0(template.name, ".tar")), package = 'ProjectTemplate')
 
 .list.files.and.dirs <- function(path) {
   # no.. not available in R 2.15.3