+++ R/default.config.R
@@ -1,7 +1,0 @@
-#' Default configuration
-#'
-#' This list stores the configuration used for missing items
-#' in the configuration of the current project.
-#'
-#' @include translate.dcf.R
-default.config <- translate.dcf(system.file('defaults/config/default.dcf', package = 'ProjectTemplate'))