+++ R/load.project.R
@@ -1,7 +1,8 @@
 #' Automatically load data and packages for a project.
 #'
 #' This function automatically load all of the data and packages used by
-#' the project from which it is called.
+#' the project from which it is called.  The behaviour can be controlled by
+#' adjusting the \code{\link{project.config}} configuration.
 #'
 #' @param override.config Named list, allows overriding individual configuration
 #'   items.
@@ -11,7 +12,7 @@
 #' @export
 #'
 #' @seealso \code{\link{create.project}}, \code{\link{get.project}},
-#'   \code{\link{cache.project}}, \code{\link{show.project}}
+#'   \code{\link{cache.project}}, \code{\link{show.project}}, \code{\link{project.config}}
 #'
 #' @examples
 #' library('ProjectTemplate')
@@ -19,24 +20,14 @@
 #' \dontrun{load.project()}
 load.project <- function(override.config = NULL)
 {
-  this_dir <- .project.info(getwd())
-  if (!this_dir$is.ProjectTemplate) {
-          return(
-          message(paste0(c(paste0("Current directory: ", basename(getwd()),
-                                  " is not a ProjectTemplate directory"),
-                         "Please change to correct directory and re-run load.project()"),
-                         collapse = "\n")
-                  )
-          )
-  }
-        
+  project_name <- .stopifnotproject("Please change to correct directory and re-run load.project()")
+
   my.project.info <- list()
 
-  message(paste0('Project name: ', this_dir$name))
+  message('Project name: ', project_name)
   message('Loading project configuration')
 
   config <- .load.config(override.config)
-
   .check.version(config)
 
   assign('config', config, envir = .TargetEnv)
@@ -44,44 +35,74 @@ load.project <- function(override.config = NULL)
 
   options(stringsAsFactors = config$as_factors)
 
-  if (config$load_libraries)
-  {
-    message('Autoloading packages')
-    my.project.info$packages <- c()
-    for (package.to.load in strsplit(config$libraries, '\\s*,\\s*')[[1]])
-    {
-      message(paste(' Loading package:', package.to.load))
-      require.package(package.to.load)
-      my.project.info$packages <- c(my.project.info$packages, package.to.load)
-    }
+  if (config$load_libraries) {
+    my.project.info <- .load.libraries(config, my.project.info)
   }
 
-  if (config$logging)
-  {
-    message('Initializing logger')
-    require.package('log4r')
-
-    logger <- log4r::create.logger()
-    .provide.directory('logs')
-
-    log4r::logfile(logger) <- file.path('logs', 'project.log')
-    log4r::level(logger) <- config$logging_level
-    assign('logger', logger, envir = .TargetEnv)
+  if (config$logging) {
+    my.project.info <- .init.logger(config, my.project.info)
   }
 
-  if (file.exists('lib'))
-  {
+  my.project.info <- .load.helpers(config, my.project.info)
+
+  if (config$data_loading | config$cache_loading) {
+    my.project.info <- .load.data(config, my.project.info)
+  }
+
+  if (config$munging) {
+    my.project.info <- .munge.data(config, my.project.info)
+  }
+
+  assign('project.info', my.project.info, envir = .TargetEnv)
+}
+
+.unload.project <- function() {
+  suppressWarnings(rm(list = c("config", "logger", "project.info"),
+                      envir = .TargetEnv))
+}
+
+## Load libraries listed in configuration into memory ------------------------
+.load.libraries <- function(config, my.project.info) {
+  message('Autoloading packages')
+  my.project.info$packages <- c()
+
+  for (package.to.load in strsplit(config$libraries, '\\s*,\\s*')[[1]]) {
+    message(' Loading package: ', package.to.load)
+    require.package(package.to.load)
+    my.project.info$packages <- c(my.project.info$packages, package.to.load)
+  }
+
+  return(my.project.info)
+}
+
+## Initialize logging through log4r package ----------------------------------
+.init.logger <- function(config, my.project.info) {
+  message('Initializing logger')
+  require.package('log4r')
+
+  logger <- log4r::create.logger()
+  .provide.directory('logs')
+
+  log4r::logfile(logger) <- file.path('logs', 'project.log')
+  log4r::level(logger) <- config$logging_level
+  assign('logger', logger, envir = .TargetEnv)
+  return(my.project.info)
+}
+
+## Load helper functions -----------------------------------------------------
+.load.helpers <- function(config, my.project.info) {
+  if (file.exists('lib')) {
     message('Autoloading helper functions')
 
     my.project.info$helpers <- c()
 
     helpers <- dir('lib', pattern = '[.][rR]$')
-    
+
     # force globals.R to be read first, if it exists
     if ("globals.R" %in% helpers) {
-            helpers<-c("globals.R", helpers[!(helpers %in% "globals.R")])
+      helpers <- c("globals.R", helpers[!(helpers %in% "globals.R")])
     }
-    
+
     deprecated.files <- intersect(
       helpers, c('boot.R', 'load_data.R', 'load_libraries.R',
                  'preprocess_data.R', 'run_tests.R'))
@@ -90,170 +111,129 @@ load.project <- function(override.config = NULL)
                     paste(deprecated.files, collapse = ', ')))
     }
 
-    for (helper.script in helpers)
-    {
-      message(paste(' Running helper script:', helper.script))
-      source(file.path('lib', helper.script))
+    for (helper.script in helpers) {
+      message(' Running helper script: ', helper.script)
+      source(file.path('lib', helper.script), local = .TargetEnv)
       my.project.info$helpers <- c(my.project.info$helpers, helper.script)
     }
   }
-
-  # First, we load everything out of cache/.
-  if (config$cache_loading)
-  {
-    message('Autoloading cache')
-
-    my.project.info$cache <- .load.cache()
-  }
-
-  # Then we consider loading things from data/.
-  if (config$data_loading)
-  {
-    message('Autoloading data')
-
-    my.project.info$data <- .load.data(config$recursive_loading)
-  }
-
-  if (config$data_tables)
-  {
-    require.package('data.table')
-
-    message('Converting data.frames to data.tables')
-
-    .convert.to.data.table(my.project.info$data)
-  }
-
-  if (config$munging)
-  {
-    message('Munging data')
-    for (preprocessing.script in sort(dir('munge', pattern = '[.][rR]$')))
-    {
-      message(paste(' Running preprocessing script:', preprocessing.script))
-      source(file.path('munge', preprocessing.script))
-    }
-  }
-
-  assign('project.info', my.project.info, envir = .TargetEnv)
-  #assign('project.info', my.project.info, envir = parent.frame())
-  #assign('project.info', my.project.info, envir = environment(create.project))
+  return(my.project.info)
 }
 
-.unload.project <- function() {
-  suppressWarnings(rm(list = c("config", "logger", "project.info"), envir = .TargetEnv))
-}
+## Load data into memory from cache/ and data/ -------------------------------
+.load.data <- function(config, my.project.info) {
+  message('Autoloading data')
 
-.normalize.config <- function(config, names, norm.fun) {
-  config[names] <- lapply(config[names], norm.fun)
-  config
-}
-
-.boolean.cfg <- function(x) {
-  ret <- if (x == 'on') TRUE
-  else if (x == 'off') FALSE
-  else as.logical(x)
-  if (is.na(ret)) stop('Cannot convert ', x, ' to logical value.')
-  ret
-}
-
-.load.cache <- function() {
-  .provide.directory('cache')
-  cache.files <- dir('cache')
-  cached.files <- c()
-
-  for (cache.file in cache.files)
-  {
-    filename <- file.path('cache', cache.file)
-
-    for (extension in ls(extensions.dispatch.table))
-    {
-      if (grepl(extension, cache.file, ignore.case = TRUE, perl = TRUE))
-      {
-        variable.name <- clean.variable.name(sub(extension,
-                                                 '',
-                                                 cache.file,
-                                                 ignore.case = TRUE,
-                                                 perl = TRUE))
-
-        # If this variable already exists in the global environment, don't load it from cache.
-        if (variable.name %in% ls(envir = .TargetEnv))
-        {
-          next()
-        }
-
-        message(paste(" Loading cached data set: ", variable.name, sep = ''))
-
-        do.call(extensions.dispatch.table[[extension]],
-                list(cache.file,
-                     filename,
-                     variable.name))
-
-        cached.files <- c(cached.files, variable.name)
-
-        break()
-      }
-    }
-  }
-
-  cached.files
-}
-
-.load.data <- function(recursive) {
-  .provide.directory('data')
-  data.files <- dir('data', recursive = recursive)
   data.files.loaded <- c()
+  cache.files.loaded <- c()
 
-  for (data.file in data.files)
-  {
-    filename <- file.path('data', data.file)
+  # List all available data
+  data.files <- .list.data(config)
+  # Order the data such that cached only variables are loaded first, and then
+  # everything in alphabetical order
+  load.order <- order(data.files$cache_only,
+                      data.files$filename,
+                      data.files$varname,
+                      decreasing = c(TRUE, FALSE, FALSE),
+                      method = "radix")
 
-    for (extension in ls(extensions.dispatch.table))
-    {
-      if (grepl(extension, data.file, ignore.case = TRUE, perl = TRUE))
-      {
-        variable.name <- clean.variable.name(sub(extension,
-                                                 '',
-                                                 data.file,
-                                                 ignore.case = TRUE,
-                                                 perl = TRUE))
+  # Loop over all rows in the list in the determined order
+  for (f in load.order) {
+    data.file <- data.files[f,]
+    variable <- data.file$varname
 
-        # If this variable already exists in cache, don't load it from data.
-        if (variable.name %in% ls(envir = .TargetEnv))
-        {
-          next()
-        }
+    # Check if file must be loaded
+    is_loaded <- variable %in% ls(envir = .TargetEnv)
+    if (is_loaded | data.file$is_ignored) {
+      next()
+    }
 
-        message(paste(" Loading data set: ", variable.name, sep = ''))
-
-        do.call(extensions.dispatch.table[[extension]],
-                list(data.file,
-                     filename,
-                     variable.name))
-
-        data.files.loaded <- c(data.files.loaded, variable.name)
-
-        break()
+    if (config$cache_loading & data.file$is_cached) {
+      # Load data from cache/
+      message(" Loading cached data set: ", variable)
+      cache.file <- file.path(.cache.dir, paste0(variable, .cache.file.ext))
+      load(cache.file, envir = .TargetEnv)
+      cache.files.loaded <- c(cache.files.loaded, variable)
+    } else if (config$data_loading) {
+      # Check if a reader was found for the file
+      has_reader <- ((class(data.file$reader[[1]]) == "function") || !(data.file$reader[[1]] == ''))
+      if (!has_reader) {
+        next()
       }
+      # Load data from data/
+      message(" Loading data set: ", variable)
+      reader.args <- list(data.file$filename,
+                          file.path('data', data.file$filename),
+                          variable)
+
+      # Register current variables
+      vars.old <- .var.diff.from()
+      # Actually load the data
+      do.call(data.file$reader[[1]], reader.args)
+      # Get new variables introduced by the reader
+      vars.new <- .var.diff.from(vars.old)
+
+      if (config$data_tables) {
+        .convert.to.data.table(vars.new)
+      } else {
+        .convert.to.tibble(vars.new)
+      }
+
+      if (config$cache_loaded_data && length(vars.new) > 0) {
+        sapply(vars.new, cache)
+      }
+
+      data.files.loaded <- c(data.files.loaded, vars.new)
     }
   }
 
-  data.files.loaded
+  my.project.info$cache <- cache.files.loaded
+  my.project.info$data <- data.files.loaded
+  return(my.project.info)
 }
 
+## Convert datasets to data.table
 .convert.to.data.table <- function(data.sets) {
   .require.package("data.table")
 
-  for (data.set in data.sets)
-  {
-    if (all(class(get(data.set, envir = .TargetEnv, inherits = FALSE)) == 'data.frame'))
-    {
-      message(paste(' Translating data.frame:', data.set))
-      assign(data.set,
-             data.table::data.table(get(data.set, envir = .TargetEnv, inherits = FALSE)),
-             envir = .TargetEnv)
+  for (data.set in data.sets) {
+    # Get current version of the dataset
+    loaded.data <- get(data.set, envir = .TargetEnv, inherits = FALSE)
+    if (all(class(loaded.data) == 'data.frame')) {
+      message(' Translating data.frame to data.table: ', data.set)
+      assign(data.set, data.table::data.table(loaded.data), envir = .TargetEnv)
     }
   }
 }
 
+## Convert datasets to tibble
+.convert.to.tibble <- function(data.sets) {
+  require.package("tibble", FALSE)
+
+  for (data.set in data.sets) {
+    # Get current version of the dataset
+    loaded.data <- get(data.set, envir = .TargetEnv, inherits = FALSE)
+    if (all(class(loaded.data) == 'data.frame')) {
+      message(' Translating data.frame to tibble: ', data.set)
+      assign(data.set, tibble::as_tibble(loaded.data), envir = .TargetEnv)
+    }
+  }
+}
+
+## Run all scripts in the munge/ directory -----------------------------------
+.munge.data <- function(config, my.project.info) {
+  message('Munging data')
+  for (preprocessing.script in sort(dir('munge', pattern = '[.][rR]$')))
+  {
+    message(' Running preprocessing script: ', preprocessing.script)
+    source(file.path('munge', preprocessing.script), local = .TargetEnv)
+  }
+  return(my.project.info)
+}
+
+# Auxiliary functions for loading/unloading projects -------------------------
+
+## Function to create directory if it doesn't exist yet
 .provide.directory <- function(name) {
   is.dir <- file.info(name)$isdir
   if (is.na(is.dir) || !is.dir) {
@@ -262,42 +242,7 @@ load.project <- function(override.config = NULL)
   }
 }
 
-.load.config <- function(override.config = NULL) {
-  config.path <- file.path('config', 'global.dcf')
-  config <- if (file.exists(config.path)) {
-    translate.dcf(config.path)
-  } else {
-    warning('You are missing a configuration file: ', config.path, ' . Defaults will be used.')
-    default.config
-  }
-
-  missing.entries <- setdiff(names(default.config), names(config))
-  if (length(missing.entries) > 0) {
-    warning('Your configuration file is missing the following entries: ',
-            paste(missing.entries, collapse = ', '), '. Defaults will be used.')
-    config[missing.entries] <- default.config[missing.entries]
-  }
-
-  if (length(override.config) > 0) {
-    config[names(override.config)] <- override.config
-  }
-
-  extra.entries <- setdiff(names(config), names(default.config))
-  extra.entries <- grep("^[^#]", extra.entries, value = TRUE)
-  if (length(extra.entries) > 0) {
-    warning('Your configuration contains the following unused entries: ',
-            paste(extra.entries, collapse = ', '), '. These will be ignored.')
-    config[extra.entries] <- NULL
-  }
-
-  config <- .normalize.config(config,
-                              setdiff(names(default.config), c("version", "libraries", "logging_level")),
-                              .boolean.cfg)
-  
-
-  config
-}
-
+## Function to check the project version against the package version
 #' @importFrom utils compareVersion
 .check.version <- function(config, warn.migrate = TRUE) {
   package.version <- .package.version()
@@ -320,19 +265,13 @@ load.project <- function(override.config = NULL)
   as.character(read.dcf(system.file("DESCRIPTION", package = "ProjectTemplate"), fields = "Version"))
 }
 
-# files that determine whether a directory is a ProjectTemplate project
-.mandatory.files <- c("config/global.dcf", "cache", "data")
+# Compare the variables (excluding functions) in the global env with a passed
+# in string of names and return the difference
+.var.diff.from <- function(given.var.list="", env=.TargetEnv) {
+  # Get variables in target environment of determine if they are a function
+  current.var.list <- sapply(ls(envir = env), function(x) is.function(get(x)))
+  current.var.list <- names(current.var.list[current.var.list == FALSE])
 
-# List of information about a project
-.project.info <- function (path=getwd()) {
-        is.ProjectTemplate <- .is.ProjectTemplate(path)
-        name <- ifelse(is.ProjectTemplate, basename(path), "")
-        list(is.ProjectTemplate=is.ProjectTemplate, name=name) 
-}
-
-# Test whether a given path is a ProjectTemplate project
-.is.ProjectTemplate <- function (path=getwd()) {
-        check_files <- file.path(path, .mandatory.files)
-        if(sum(file.exists(check_files))==length(check_files)) return(TRUE)
-        return(FALSE)
+  # return those not in list
+  setdiff(current.var.list, given.var.list)
 }