+++ R/translate.dcf.R
@@ -3,6 +3,10 @@
 #' This function will read a DCF file and translate the resulting
 #' data frame into a list. The DCF format is used throughout ProjectTemplate
 #' for configuration settings and ad hoc file format specifications.
+#' 
+#' The content of the DCF file are stored as character strings.  If the content
+#' is placed between the back tick character , then the content is 
+#' evaluated as R code and the result returned in a string
 #'
 #' @param filename A character vector specifying the DCF file to be
 #'   translated.
@@ -19,5 +23,16 @@
 translate.dcf <- function(filename)
 {
   settings <- read.dcf(filename)
-  setNames(as.list(as.character(settings)), colnames(settings))
+  settings <- setNames(as.list(as.character(settings)), colnames(settings))
+  
+  # Check each setting to see if it contains R code
+  for (s in names(settings)) {
+          value <- settings[[s]]
+          r_code <- gsub("^`(.*)`$", "\\1", value)
+          if (nchar(r_code) != nchar(value)) {
+                  settings[[s]] <- eval(parse(text=r_code))
+          }
+  }
+  settings
 }
+